% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_plot.R
\name{QQ_plot}
\alias{QQ_plot}
\title{interRate}
\usage{
QQ_plot(ress)
}
\arguments{
\item{data}{list of the standardized residuals.}
}
\description{
QQ plot of the standardized residuals.
}
\examples{
QQ_plot(df)
}
